services/hrtaj_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pandas==2.2.3
openpyxl==3.1.5
supabase==2.10.0
python-dotenv==1.0.1
tenacity==8.3.0
httpx==0.27.0
pytest==8.3.3
